requirements.txt
ansicolors==1.1.8
asgiref==3.4.1
astroid==2.5
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
distlib==0.3.3
filelock==3.3.0
flake8==3.8.4
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Login==0.5.0
h11==0.12.0
identify==2.3.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
jwt==1.3.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==0.13.1
pre-commit==2.9.3
py==1.10.0
py-healthcheck==1.10.1
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.2.0
PyJWT==2.0.1
pylint==2.7.1
pyparsing==2.4.7
pytest==6.2.1
python-dotenv==0.15.0
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
rfc3339==6.2
six==1.16.0
SQLAlchemy==1.3.22
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.14.0
virtualenv==20.8.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.6.0
